Added a /host/ URL for the person hosting the game.
Reason: The ID is generated on page refresh, so if a play URL
is refreshed, both the HOST and the JOIN will have different IDs
meaning both try to start the game.

Instead the HOST route resets window.ourID to the ID in the URL
so that the HOST plays properly and waits for the JOINer to init the game.