Added game over communication, sync and checking
Both players are *ALWAYS* checking for game state.
The one who JOINS always initiates the sync, by sending
its state using a 'gameover' action. The host receives this,
checks that it tallies with its own game state, and sends back confirmation.
The JOINer then uses this confirmation and updates its own state (not the model
state, which is already updated, but the general game state)