Make game resettable on refresh
If joiner refreshes halfway through the game, both clients sync back to start.
All DOM elements of the grid and hud are now created on the fly.

GridModel.defaults is a function returning an object so that issues with
cloning and deep references don't affect the model reset. Defaults are shared
among models.